requirements.txt
PyMuPDF
pandas
pillow
tqdm
openai
requests